Compile and link to ST library.
Completely change the approach of the library. Rather than using the
SparkFun approach of replaying i2c commands, instead let's integrate
with the ST C library and build the Rust FFI interface on top.

Now that the library is more robust and reliable, it makes sense to
publish it on cargo as v1.0.0.